Restructure Output controller by segregating Reconciler
